Add data binding test
This test is testing that auto timestamp properties are by default
excluded from mass property binding.  We have unit tests in grails-core
for this and those tests work on classes that are explicitly marked
with @Entity in the source code.  This test is covering a problem that
relates to indetifying a domain class as being a domain class if it is
used as a command object.  Testing this in a real app (like this one)
covers that problem.